feat:badge component, 정정 신청 내역 페이지 css 수정 -#35
