+++ src/pages/salary-correction/components/StatusBadge.tsx
@@ -1,0 +1,10 @@
+import { Badge } from "./StatusBadge.styled";
+
+interface BadgeProps {
+	status: string;
+}
+const StatusBadge = ({ status }: BadgeProps) => {
+	return <Badge $status={status}>{status}</Badge>;
+};
+
+export default StatusBadge;
